pass a targetType to the findTerm request; reactivate nav integration tests
